+++ CHANGELOG.md
@@ -25,7 +25,7 @@ for now removed features.
 
 ### Fixed
 
-for any bug fixes.
+- **dropdown**: if the user presses an arrow (up or down) on the keyboard, the browser is prevented from scrolling
 
 ### Security
 
@@ -40,7 +40,7 @@ in case of vulnerabilities.
 ### Fixed
 
 - **dropdown**: add keyboard navigation support
-- **button**: Passes the call of the `focus()` method on the web component to the inner native button.
+- **button**: passes the call of the `focus()` method on the web component to the inner native button.
 - **grid**: fix pagination to never have a negative current page
 - **tabs**: an error was raised when the component tried to access the tabs if the list of slotted tab was empty (for
   example when the component is not yet rendered).