+++ CHANGELOG.md
@@ -9,12 +9,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- **formly-field-tooltip**: add `interactive` in templateOptions
-- **formly-field-tooltip**: add `dynamicContent` in templateOptions. This will allow to update the tooltip content dynamically (using `expressionProperties` for instance). Use the default `content` templateOptions if the tooltip content is static. 
+for new features.
 
 ### Changed
 
-- **accordion-header**: aligns the header text to the left
+for changes in existing functionality.
 
 ### Deprecated
 
@@ -26,15 +25,30 @@ for now removed features.
 
 ### Fixed
 
+for any bug fixes.
+
+### Security
+
+in case of vulnerabilities.
+
+## [2.17.0] - 2022-10-18
+
+### Added
+
+- **formly-field-tooltip**: add `interactive` in templateOptions
+- **formly-field-tooltip**: add `dynamicContent` in templateOptions. This will allow to update the tooltip content dynamically (using `expressionProperties` for instance). Use the default `content` templateOptions if the tooltip content is static. 
+
+### Changed
+
+- **accordion-header**: aligns the header text to the left
+
+### Fixed
+
 - **dropdown**: if the user presses an arrow (up or down) on the keyboard, the browser is prevented from scrolling
 - **docs**: made some corrections on the [Framework integration documentation](https://sncf.gitlab.io/wcs/develop/?path=/story/documentation-framework-integrations--page)
 - **checkbox**: label does not overlap anymore on very narrow containers
 - **tooltip**: fix tooltip appearance when it is contained into a fullscreen element
 - **label**: changed selector and disposition to inline-block for several lines field label
-
-### Security
-
-in case of vulnerabilities.
 
 ## [2.16.0] - 2022-09-27